NN: One single build run now generates both i386 and amd64
git-svn-id: http://svn.zentyal.org/zentyal/trunk@19098 866bc4c1-54da-0310-a958-9d5a9a0e8d04

+++ extra/installer/generate_extras.sh
@@ -3,6 +3,17 @@
 test -r build_cd.conf || exit 1
 . ./build_cd.conf
 
+ARCH=$1
+
+if [ "$ARCH" -ne "i386" -o "$ARCH" -ne "amd64" ]
+then
+    echo "Usage: $0 [i386|amd64]"
+    exit 1
+fi
+
+EXTRAS_DIR="$EXTRAS_DIR_BASE-$ARCH"
+CHROOT="$CHROOT_BASE-$ARCH"
+
 test -d $EXTRAS_DIR || mkdir $EXTRAS_DIR
 
 sudo rm -fr $CHROOT
